finish tests and save treasury
