+++ .phpcs.xml.dist
@@ -30,6 +30,9 @@
 	<rule ref="WordPress">
 		<!-- Allow short array syntax [] -->
 		<exclude name="Generic.Arrays.DisallowShortArraySyntax"/>
+		<!-- Exclude filename rules for Plugin_Main.php -->
+		<exclude name="WordPress.Files.FileName.NotHyphenatedLowercase"/>
+		<exclude name="WordPress.Files.FileName.InvalidClassFileName"/>
 	</rule>
 	<rule ref="WordPress.NamingConventions.PrefixAllGlobals">
 		<properties>
